requirements.txt
pynput==1.7.6
pillow==10.3.0
